Some rules now...
The game will quit if you hit a bomb now...
If you mark all of the mines correctly you win!

I would like to add a summary printout for after you win/lose,
but there's no time for that now. I also need to add something
at the top of the window for status while you're playing, but
that'll have to come later.