Started fixing recursion megadeath
Beginning with the spaghetti code I wrote late at night...
    * The add_cell function was terrible, but thankfully is now less terrible
    * The destroy_cell function was, and still is, terrible. It leaks

The problem that I lazily tried to fix is that the head pointer
is updated in the function, but not returned. I need to make it a double
so bad crap stops happening.

I also remembered that ncurses is leaky...