It's been a while...
Long enough that I totally forgot what was wrong with this.
Thanks, summer job.

The problem seems to be the destroy_cells function, which
gives an invalid read of size 8. I'm going to rewrite it
a different way to see if that fixes the problem.

I will also need to determine if there's something wrong with
the clear_area function...because the area isn't clearing.

I also updated the README with compile instructions...for myself
more than anyone else.